requirements.txt
Pillow
PyMuPDF
spacy
numpy
wordfreq
torch
